Complete homework to lesson 5
Do not realise fetch with findById to web version!

+++ GB_Spring_3/src/main/java/com/gb/spring/lessons/GB_Spring/controllers/ProductController.java
@@ -23,7 +23,7 @@ public class ProductController {
     }
 
     @PostMapping("/productInfo")
-    public String getProductInfo(@RequestParam Integer id, Model model){
+    public String getProductInfo(@RequestParam Long id, Model model){
         model.addAttribute("product", productService.getOneProduct(id));
         return  "productInfo";
     }
@@ -40,7 +40,7 @@ public class ProductController {
     }
 
     @PostMapping("/delete")
-    public String deleteProduct(@RequestParam Integer id){
+    public String deleteProduct(@RequestParam Long id){
         productService.deleteProductById(id);
         return "redirect:all";
     }
